:shirt: Move color code to namespace
