+++ .vscode/settings.json
@@ -15,6 +15,7 @@
     "prettier.proseWrap": "always",
     "cSpell.words": [
         "Argb",
+        "Rgba",
         "Texim"
     ],
 }
