+++ src/Texim.Tool/BlackRockShooter/Binary2Ptmd.cs
@@ -27,6 +27,7 @@ namespace Texim.BlackRockShooter
 {
     using System;
     using System.Drawing;
+    using Texim.Colors;
     using Texim.Palettes;
     using Yarhl.FileFormat;
     using Yarhl.IO;
