+++ src/Texim.Tool/Disgaea/Binary2Ykcmp.cs
@@ -20,6 +20,7 @@
 namespace Texim.Disgaea
 {
     using System;
+    using Texim.Colors;
     using Texim.Palettes;
     using Yarhl.FileFormat;
     using Yarhl.IO;
