+++ src/Texim/Colors/IOExtensions.cs
@@ -17,22 +17,39 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
-namespace Texim.Palettes
+namespace Texim.Colors
 {
-    using System;
     using System.Collections.Generic;
-    using System.IO;
+    using Yarhl.IO;
 
-    public interface IColorEncoding
+    public static class IOExtensions
     {
-        Rgb Decode(Stream stream);
+        public static Rgb ReadColor<T>(this DataReader reader)
+            where T : IColorEncoding, new()
+        {
+            T encoding = new T();
+            return encoding.Decode(reader.Stream);
+        }
 
-        Rgb[] Decode(Stream stream, int numColors);
+        public static Rgb[] ReadColors<T>(this DataReader reader, int numColors)
+            where T : IColorEncoding, new()
+        {
+            T encoding = new T();
+            return encoding.Decode(reader.Stream, numColors);
+        }
 
-        Rgb[] Decode(Span<byte> data);
+        public static void Write<T>(this DataWriter writer, Rgb color)
+            where T : IColorEncoding, new()
+        {
+            T encoding = new T();
+            writer.Write(encoding.Encode(color));
+        }
 
-        byte[] Encode(Rgb color);
-
-        byte[] Encode(IEnumerable<Rgb> colors);
+        public static void Write<T>(this DataWriter writer, IEnumerable<Rgb> colors)
+            where T : IColorEncoding, new()
+        {
+            T encoding = new T();
+            writer.Write(encoding.Encode(colors));
+        }
     }
 }
