+++ src/Texim/Palettes/IPalette.cs
@@ -20,6 +20,7 @@
 namespace Texim.Palettes
 {
     using System.Collections.ObjectModel;
+    using Texim.Colors;
     using Yarhl.FileFormat;
 
     public interface IPalette : IFormat
