+++ src/Texim/Palettes/Palette.cs
@@ -22,6 +22,7 @@ namespace Texim.Palettes
     using System.Collections.Generic;
     using System.Collections.ObjectModel;
     using System.Linq;
+    using Texim.Colors;
 
     public class Palette : IPalette
     {
